+ added javadoc to Dataloader + added javadoc to all Relation services + added javadoc to the Model Assembler for Relation + added javadoc to all Task services + added minor pretty modification to data loader of easier output reading & more detailed information + fixed code to remove some warnings

+++ Backend-spring/src/main/java/com/todo/challengev2/relations/functionalities/delete/RelationDeleteServiceImpl.java
@@ -13,8 +13,12 @@ public class RelationDeleteServiceImpl implements RelationDeleteService{
     @Autowired
     private RelationRepository relationRepository;
 
+    /**
+     * Attempts to delete the Relaion object with the target Id from the RelationRepository
+     * @param id the id of the target object to be deleted
+     */
     @Override
     public void deleteRelation(UUID id) {
-        relationRepository.delete(relationRepository.findById(id).get());
+        if (relationRepository.findById(id).isPresent()) relationRepository.delete(relationRepository.findById(id).get());
     }
 }
